feat: allow spending specific vtxos
We extend `sendBitcoin(..)``by a new optional paramter: `selectedVtxos?: {inputs: VirtualCoin[], changeAmount: number};`.

This allows the caller to spend sepcific VTXOs.

Note: If `params.amount` equals the sum of all selected VTXOs, the remainder will be returned as change output. If the sum of all selected VTXOs does not cover `params.amount` an error is returned.